requirements.txt
# Core
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data
yfinance>=0.2.35
jugaad-data>=0.24
nsetools>=1.0.11
pandas-ta>=0.3.14b0
# ta-lib>=0.4.28  # Optional: If ta-lib not available, pandas-ta is used as fallback

# Model
scipy>=1.10.0

# Hyperopt
optuna>=3.4.0

# Evaluation
pyarrow>=12.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utils
tqdm>=4.65.0
pyyaml>=6.0
loguru>=0.7.0

# Development
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
